drop component install for cpack DEB packages
Component package creation for DEB packages in cpack is not suitable:
packages get wrong names, one cannot add extra stuff to compnent
groups, etc.

As long as cpack does not get extra features or revamps this we
can't use it as our package creation strategy. For now, we place
all we have into a single package - YES, this includes dev-files
in the core package - and got with this until we find a better
way to handle core and dev package creation better. =(