omit message flow particle debug output in general
The observed MOD-SEND and MOD-RECV debug messages tend to pollute
the debug output. So a dedicated property of modules has been
created:    <bool> debug_message_flow

If turned on all module's send and recv actions are protocolled on
stderr. If turned off (the default) the module does not state anything
about received or sent messages.

This flag is independent of the normal debug flag.