qkd-resize does now concatenate disclosed keys
Previously the key material of disclosed keys were rejected
by the qkd-resize module. After a long discussion with Christoph
Pacher one is on the better, safe side to include the disclosed
key in the qkd-resize'd key.

Up to this commit qkd-resize concatenated the QBER and the
disclosed bits on every key it consumed. In same cases keys
had to be fully disclosed. The metadata for such keys, like the
QBER and the number of disclosed bits (which is 100% of the key)
have been added to the new resized key. However the key material
of the disclosed keys have been spared.

This changed now to include also the disclosed data in the
resulting key material.