adds automatic routing support into the node
This is a big one. With this the q3pd node understands
the "nic = " option per link in the config file.

    nic = LOCAL-IP to REMOTE-IP

Example:

    [link.alice_to_bob]
    nic = 10.81.39.121 to 10.81.39.122

When the Q3P link has established a connection the local
network interface card ("nic") will be raised. Depending
on the current operating system state, it will be labeled
"q3p0", "q3p1", "q3p2",...

The above configuration line tells the link to attach the
IPv4 address "10.81.39.121" to its local side of the link.
Furthermore the link also adds a route to the kernel routing
tables: 10.81.39.122 via 10.81.38.121.

This can be run as user if the binary (the q3pd program) has
the network admin capabilities set (see setcap and getcap
superuser commands).