sets default PA values to ident
The PA works with both security bits and reduction rate. This commit
sets the security bits to 0 and the reduction rate to 1.0 which renders
a key of identical length (though this is not recommended for
production).

Oh, and BTW, some code clean up in the PA code as well. While I'm at it.