introduces security bits XOR reduction rate for PA
Again, security bits and reduction rate is exclusive in PA.
If security bits are set then the next key size will be
calculated with security bits. If the user sets the reduction
rate then only the reduction rate is used. To figure out which
one is currently used the PA object has a new DBus property
"calculation_procedure". If this is 0 then security bits are
used, if this is 1 then reduction rate is used.