enforces key-ids sent and received
On peer to peer module communication this commit enforces the
key id to be sent along. This is one step to create a better
key sync since the module can now verify, if the peer module
is working on the very same key as itself.

If a key mismatch is detected a qkd::exception::protocol_error
exception is thrown.